rename internal/common to internal/pathtools
