+++ internal/packages/BUILD.bazel
@@ -13,8 +13,8 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/internal/packages",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/common:go_default_library",
         "//internal/config:go_default_library",
+        "//internal/pathtools:go_default_library",
         "@com_github_bazelbuild_buildtools//build:go_default_library",
     ],
 )
