+++ internal/packages/walk.go
@@ -24,8 +24,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/bazelbuild/bazel-gazelle/internal/common"
 	"github.com/bazelbuild/bazel-gazelle/internal/config"
+	"github.com/bazelbuild/bazel-gazelle/internal/pathtools"
 	bf "github.com/bazelbuild/buildtools/build"
 )
 
@@ -90,7 +90,7 @@ func Walk(c *config.Config, root string, f WalkFunc) {
 		// Check if this directory should be updated.
 		if !isUpdateDir {
 			for _, updateRel := range updateRels {
-				if common.PathHasPrefix(rel, updateRel) {
+				if pathtools.HasPrefix(rel, updateRel) {
 					isUpdateDir = true
 				}
 			}
