+++ internal/pathtools/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["path.go"],
-    importpath = "github.com/bazelbuild/bazel-gazelle/internal/common",
+    importpath = "github.com/bazelbuild/bazel-gazelle/internal/pathtools",
     visibility = ["//:__subpackages__"],
 )
 
@@ -11,5 +11,5 @@ go_test(
     name = "go_default_test",
     srcs = ["path_test.go"],
     embed = [":go_default_library"],
-    importpath = "github.com/bazelbuild/bazel-gazelle/internal/common",
+    importpath = "github.com/bazelbuild/bazel-gazelle/internal/pathtools",
 )
