+++ internal/resolve/BUILD.bazel
@@ -14,8 +14,8 @@ go_library(
     importpath = "github.com/bazelbuild/bazel-gazelle/internal/resolve",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/common:go_default_library",
         "//internal/config:go_default_library",
+        "//internal/pathtools:go_default_library",
         "@com_github_bazelbuild_buildtools//build:go_default_library",
         "@org_golang_x_tools//go/vcs:go_default_library",
     ],
