+++ internal/resolve/label.go
@@ -22,7 +22,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/bazelbuild/bazel-gazelle/internal/common"
+	"github.com/bazelbuild/bazel-gazelle/internal/pathtools"
 )
 
 // A Label represents a label of a build target in Bazel.
@@ -134,5 +134,5 @@ func packageContains(repo, pkg string, label Label) bool {
 	if label.Relative {
 		log.Panicf("label must not be relative: %s", label)
 	}
-	return repo == label.Repo && common.PathHasPrefix(label.Pkg, pkg)
+	return repo == label.Repo && pathtools.HasPrefix(label.Pkg, pkg)
 }
