+++ internal/resolve/resolve.go
@@ -22,8 +22,8 @@ import (
 	"path"
 	"strings"
 
-	"github.com/bazelbuild/bazel-gazelle/internal/common"
 	"github.com/bazelbuild/bazel-gazelle/internal/config"
+	"github.com/bazelbuild/bazel-gazelle/internal/pathtools"
 	bf "github.com/bazelbuild/buildtools/build"
 )
 
@@ -232,8 +232,8 @@ func (r *Resolver) resolveGo(imp string, from Label) (Label, error) {
 		return label, nil
 	}
 
-	if common.PathHasPrefix(imp, r.c.GoPrefix) {
-		return r.l.LibraryLabel(common.PathTrimPrefix(imp, r.c.GoPrefix)), nil
+	if pathtools.HasPrefix(imp, r.c.GoPrefix) {
+		return r.l.LibraryLabel(pathtools.TrimPrefix(imp, r.c.GoPrefix)), nil
 	}
 
 	return r.external.resolve(imp)
