+++ internal/resolve/resolve_external.go
@@ -21,7 +21,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/bazelbuild/bazel-gazelle/internal/common"
+	"github.com/bazelbuild/bazel-gazelle/internal/pathtools"
 	"golang.org/x/tools/go/vcs"
 )
 
@@ -79,7 +79,7 @@ func (r *externalResolver) resolve(importpath string) (Label, error) {
 
 	var pkg string
 	if importpath != prefix {
-		pkg = common.PathTrimPrefix(importpath, prefix)
+		pkg = pathtools.TrimPrefix(importpath, prefix)
 	}
 
 	label := r.l.LibraryLabel(pkg)
